Updated the base functions to account for Windows systems
